split police
