--- include/threadpool/police/scheduling.hpp
@@ -1,183 +1,0 @@
-#pragma once
-#include <functional>
-#include <mutex>
-#include <queue>
-#include <random>
-
-namespace thread_pool
-{
-	template <typename _Task>
-	class priority_scheduler
-	{
-	public:
-		bool empty()
-		{
-			return queue_.empty();
-		}
-
-		void push(const _Task& task)
-		{
-			queue_.push(task);
-		}
-
-		bool pop(_Task& t)
-		{
-			if (queue_.empty())
-				return false;
-
-			t = queue_.top();
-
-			queue_.pop();
-
-			return true;
-		}
-
-		void clear()
-		{
-			while (!queue_.empty())
-			{
-				queue_.pop();
-			}
-		}
-
-		std::size_t size()
-		{
-			return queue_.size();
-		}
-
-	private:
-		std::priority_queue<_Task> queue_;
-	};
-
-	template <typename _Task>
-	class fifo_scheduler
-	{
-	public:
-		void push(_Task&& t)
-		{
-			queue_.push(std::forward<_Task>(t));
-		}
-
-		bool pop(_Task& t)
-		{
-			if (queue_.empty())
-				return false;
-
-			t = std::move(queue_.front());
-
-			queue_.pop();
-
-			return true;
-		}
-
-		void clear()
-		{
-			while (!queue_.empty())
-			{
-				queue_.pop();
-			}
-		}
-
-		std::size_t size()
-		{
-			return queue_.size();
-		}
-
-		bool empty()
-		{
-			return queue_.empty();
-		}
-
-	private:
-		std::queue<_Task> queue_;
-	};
-
-	template <typename _Task>
-	class multi_fifo_schedule
-	{
-	public:
-		multi_fifo_schedule(std::size_t queue_size = std::thread::hardware_concurrency())
-			: queues_(queue_size)
-		{}
-
-		void push(_Task&& t)
-		{
-			auto index = find_less();
-
-			queues_[index].push(std::forward<_Task>(t));
-		}
-
-		bool pop(_Task& t)
-		{
-			std::random_device rd;
-			std::mt19937 gen(rd());
-
-			std::uniform_int_distribution<> dis(0, static_cast<int>(queues_.size() - 1));
-
-			int id = dis(gen);
-
-			while (queues_.at(id).empty())
-			{
-				++id > queues_.size() - 1 ? id = 0 : 0;
-			}
-
-			t = queues_.at(id).front();
-
-			queues_.at(id).pop();
-
-			return true;
-		}
-
-		bool empty()
-		{
-			return std::all_of(queues_.begin(), queues_.end(), [](auto queue) { return queue.empty(); });
-		}
-
-		std::size_t size()
-		{
-			return queues_.size();
-		}
-
-	private:
-		std::size_t find_less()
-		{
-			std::size_t total{};
-			int index = 0;
-
-			for (std::size_t i = 0; i < queues_.size(); ++i)
-			{
-				if (queues_[i].size() > total)
-					continue;
-
-				total = queues_[i].size();
-
-				index = i;
-			}
-
-			return index;
-		}
-
-		std::size_t find_large()
-		{
-			std::size_t total{};
-			int index = 0;
-
-			for (std::size_t i = 0; i < queues_.size(); ++i)
-			{
-				if (queues_[i].size() < total)
-					continue;
-
-				total = queues_[i].size();
-
-				index = i;
-			}
-
-			return index;
-		}
-
-	private:
-		std::vector<std::queue<_Task>> queues_;
-
-		std::vector<std::string> real_list_;
-	};
-} // namespace thread_pool
